fix(code standards): bring over more linters and the github integration from hub repo (#865)

+++ types/validation.go
@@ -51,7 +51,7 @@ func (b *Block) ValidateWithState(state State) error {
 	if state.LastBlockHeight <= 0 && b.Header.Height != uint64(state.InitialHeight) {
 		return errors.New("initial b height mismatch")
 	}
-	if state.LastBlockHeight > 0 && b.Header.Height != uint64(state.LastStoreHeight)+1 {
+	if state.LastBlockHeight > 0 && b.Header.Height != state.LastStoreHeight+1 {
 		return errors.New("b height mismatch")
 	}
 	if !bytes.Equal(b.Header.AppHash[:], state.AppHash[:]) {
